velocity pid (SEE EXTENDED COMMIT)
THIS IMPLEMENTATION MAY CAUSE THE DRIVETRAIN TO LOCK TO WHATEVER SPEED IS PASSED TO THE moveVelocityPID() METHOD UNTIL IT IS EXPLICITLY STOPPED. This means that the robot may start moving and not stop until you explicitly tell it to. DO NOT attempt to use this method unless you have a fast, reliable way to stop the runaway bot (for example, binding Drive.stop() to a button) or the bot is on blocks.

--Gabe